Complete task permissions and drag drop workflow

+++ apps/web/src/lib/date-time.ts
@@ -1,5 +1,5 @@
 // Reuse formatters; no timeZone override means the viewer's local timezone.
-const date = new Intl.DateTimeFormat("en-GB", { day: "2-digit", month: "short", year: "numeric" });
+const date = new Intl.DateTimeFormat("en-GB", { day: "2-digit", month: "long", year: "numeric" });
 const time = new Intl.DateTimeFormat("en-US", { hour: "2-digit", minute: "2-digit", hour12: true });
 export function formatLocalDateTime(value: string | null | undefined) {
   if (!value) return "Not set";
